Extract useOpenExpandedAiChat shared by the drawer item and expand button
